// apps/hanwha-ocean/src/pages/crane-detail/model/types.ts
export interface CmmsHoistUnit {
  runFault: RunFaultStatus;
  joystickStep: string;
  speedRef: number;
  actSpeed: number;
  current: number;
  driveFault: string;
  position: number;
  load: number;
  motorTemp: number;
  // 상태 비트
  mechStopRelay: OnOff;
  driveFaultBit: OnOff;
  driveMainCb: OnOff;
  driveFanCb: OnOff;
  fieldCb: OnOff;
  driveMainMc: OnOff;
  driveFanMc: OnOff;
  driveFieldMc: OnOff;
  driveFieldFuseBlown: OnOff;
  motorFanMainCb: OnOff;
  motorFanMc: OnOff;
  lubricationMotorCb: OnOff;
  eStopPb: OkNg;
}